Documentation: some method docblock format improvements
A method/function docblock has guidelines for information order and such. It is typically laid out like:
```php
/**
 * Short description.
 *
 * Long description....
 *
 * @tag Tag description.
 * @tag Tag description.
 *
 * @param type $name Parameter description.
 *
 * @return type Description.
 */
```

While the order of some of these are not 100% clearly defined, the short and long description _have_ to be at the top and the `@param` and `@return` tags have to be at the bottom and shouldn't be combined in a block with other tags.